eth and eth-contract types updated
