+++ packages/web3-eth/types/index.d.ts
@@ -33,7 +33,9 @@ import {
     Common,
     hardfork,
     chain,
-    BlockNumber
+    BlockNumber,
+    LogsOptions,
+    PastLogsOptions
 } from 'web3-core';
 import { Subscription } from 'web3-core-subscriptions';
 import { AbiCoder } from 'web3-eth-abi';
@@ -387,16 +389,6 @@ export interface BlockTransactionObject extends BlockTransactionBase {
 
 export interface BlockTransactionString extends BlockTransactionBase {
     transactions: string[];
-}
-
-export interface PastLogsOptions extends LogsOptions {
-    toBlock?: BlockNumber;
-}
-
-export interface LogsOptions {
-    fromBlock?: BlockNumber;
-    address?: string | string[];
-    topics?: Array<string | string[] | null>;
 }
 
 export interface GetProof {
